Index

PMP.PearsonType1 — Type

PearsonType1(a,b,α,β)

The Pearson Type 1 distribution with shape parameters α and β defined on the interval (a, b) has the probability density function for $a<y<b$

\[f(y; a, b, \alpha, \beta) = \frac{1}{B(\alpha, \beta)} \frac{(y-a)^{\alpha-1} (b-y)^{\beta-1}}{(b-a)^{\alpha+\beta-1}},\]

and 0 elsewhere.

PearsonType1()   # Pearson Type 1 distribution on the unit interval with shape parameters (1,1) i.e. Uniform(0, 1)

params(d)        # Get the parameters, i.e. (a, b, α, β)

location(d)      # Get the location parameter, i.e. a
scale(d)         # Get the scale parameter, i.e. b-a
shape(d)         # Get the shape parameters, i.e. (α, β)

External links

Pearson Type 1 distribution on Wikipedia

PMP.PearsonType1b — Type

PearsonType1b(b,α,β)

The Pearson Type 1 distribution with shape parameters α and β defined on the interval (0, b) has the probability density function for $a<y<b$

\[f(y; b, \alpha, \beta) = \frac{1}{B(\alpha, \beta)} \frac{y^{\alpha-1} (b-y)^{\beta-1}}{b^{\alpha+\beta-1}},\]

and 0 elsewhere.

PearsonType1b()   # Pearson Type 1 distribution on the unit interval with shape parameters (1,1) i.e. Uniform(0, 1)

params(d)        # Get the parameters, i.e. (b, α, β)

scale(d)         # Get the scale parameter, i.e. b
shape(d)         # Get the shape parameters, i.e. (α, β)

External links

Pearson Type 1 distribution on Wikipedia

PMP.PMP_GEV — Method

PMP_GEV(rain::Vector{<:Real}, date::Vector{DateTime}, return_time::Real, d₁::Int, d₂::Int)
PMP_GEV(rain::Vector{<:Real}, date::Vector{DateTime}, return_time::Real)

Estimation of the PMP by univariate GEV method, with a chosen return time.

PMP.PMP_Hershfield — Method

PMP_Hershfield(rain_daily::Vector{<:Real}, date::Vector{DateTime}, K::Real, d₁::Int, d₂::Int)
PMP_Hershfield(rain_daily::Vector{<:Real}, date::Vector{DateTime}, K::Real)
PMP_Hershfield(rain_daily::Vector{<:Real}, date::Vector{DateTime}, d₁::Int, d₂::Int)
PMP_Hershfield(rain_daily::Vector{<:Real}, date::Vector{DateTime})

Estimation of the PMP by Hershfield empirical method.

PMP.PMP_mm — Function

PMP_mm(rain_storm::Vector{<:Real}, pw_storm::Vector{<:Real}, date_storm::Vector{DateTime}, pw_max::Vector{<:Real})
PMP_mm(rain_storm::Vector{<:Real}, pw_storm::Vector{<:Real}, date_storm::Vector{Date}, pw_max::Vector{<:Real})

Estimation of the PMP by moisture maximization.

PMP.PW_max — Function

PW_max(pw::Vector{<:Real}, date::Vector{DateTime})
PW_max(pw::Vector{<:Real}, date::Vector{Date})

Estimate the maximum precipitable water for each month of interest.

PMP.PW_return_period — Function

PW_return_period(pw::Vector{<:Real}, date::Vector{DateTime}, return_period::Int=100)
PW_return_period(pw::Vector{<:Real}, date::Vector{Date}, return_period::Int=100)

Estimate the precipitable water return value for each month of interest for a given return period.

PMP.PW_storm — Function

PW_storm(storm_date::Vector{DateTime}, dewpoint::Vector{<:Real}, dewpoint_date::Vector{DateTime}, d₂::Int, frequency::Int, time_int::Int=12)
PW_storm(storm_date::Vector{Date}, dewpoint::Vector{<:Real}, dewpoint_date::Vector{DateTime}, d₂::Int, frequency::Int, time_int::Int=12)
PW_storm(storm_date::Vector{DateTime}, dewpoint::Vector{<:Real}, dewpoint_date::Vector{Date}, d₂::Int, frequency::Int, time_int::Int=12)
PW_storm(storm_date::Vector{Date}, dewpoint::Vector{<:Real}, dewpoint_date::Vector{Date}, d₂::Int, frequency::Int, time_int::Int=12)

Get the precipitable water for each storm.

PMP.dataset — Method

dataset(name::String)::DataFrame

Load the dataset associated with name.

Datasets available:

rain: observed precipitations (in mm) recorded at the Montréal-Trudeau International Airport;

dewpoint: observed dew point (in °C) recorded at the Montréal-Trudeau International Airport.

Examples

julia> PMP.dataset("rain")

PMP.dewpoint_to_PW — Method

dewpoint_to_PW(dew_data::Real)

Convert dew point observation in precipitable water (PW).

The relation is given by the Table A.1.1 of the annex of the "Manual on Estimation of Probable Maximum Precipitation (PMP)" (WMO, 2009).

PMP.get_max_persisting_dew — Function

get_max_persisting_dew(dewpoint::Vector{<:Real}, frequency::Int, time_int::Int=12)

Get the maximum persisting dewpoint of a storm for which data are taken at a given frequency.

The highest persisting dewpoint for some specified time interval is the value equalled or exceeded at all observations during the period (WMO, 2009).

PMP.storm_maximization — Function

storm_maximization(rain_storm::Vector{<:Real}, pw_storm::Vector{<:Real}, date_storm::Vector{DateTime}, pw_max::Vector{<:Real})
storm_maximization(rain_storm::Vector{<:Real}, pw_storm::Vector{<:Real}, date_storm::Vector{Date}, pw_max::Vector{<:Real})

Estimation of the maximization ratio, effective precipitation and maximized precipitation.

PMP.storm_selection — Function

storm_selection(rain::Vector{<:Real}, date::Vector{DateTime}, p::Real, d₁::Int, d₂::Int=72)
storm_selection(rain::Vector{<:Real}, date::Vector{Date}, p::Real, d₁::Int, d₂::Int=72)
storm_selection(rain::Vector{<:Real}, date::Vector{DateTime}, p::Real)
storm_selection(rain::Vector{<:Real}, date::Vector{Date}, p::Real)

Select the PMP magnitude storms of each year.

The function choose the greatest precipitations of duration d₂ from data of duration d₁ while avoiding overlap. The function then select the p (a proportion) greatest storm of each year to be maximized.

PMP.total_precipitation — Function

total_precipitation(rain::Vector{<:Real}, date::Vector{DateTime}, d₁::Int, d₂::Int)
total_precipitation(rain::Vector{<:Real}, date::Vector{Date}, d₁::Int, d₂::Int)

Estimate the greatest precipitations taken over a given duration d₁ on a longer duration d₂.

The function choose the greatest precipitations of duration d₂ while avoiding overlap. Datasets rain and date should not contain missing data.